Refactor the code and add docstring
Multiple checks are performed now.
1. Connector node’s current community nodes are checked with and without the connector node
2. Connector node’s neighboring community nodes are checked with and without the connector node.

If the sim score of the current community with connector node is lower than without connector node AND if the sim score of neighboring community with connector node is higher than without, only then the connector node is moved to the neighboring community. If in case this is not the case, the node is placed in a separate community.

Also added check for node disconnection if the connector nodes are moved.